# openapi.yaml
...
openapi: 3.1.0
info:
  title: photo-portfolio-store
  version: 0.1.0
  description: |
    Photographer portfolio + print storefront API.

    Bounded contexts: auth, events, media, search, catalog, commerce,
    compliance, admin. Each operation declares its context via tag.
servers:
  - url: http://localhost:4000
    description: local
...
security: []
paths:
...
  /v1/consents/biometric/disclosure:
    get:
      operationId: getBiometricDisclosure
      tags: [compliance]
      summary: Statutory biometric disclosure text (BIPA 15(a) / informed consent)
      description: Public read. Returns the active disclosure document the subject must read before a covered-jurisdiction enrolment proceeds.
      parameters:
        - name: jurisdiction
          in: query
          required: false
          schema:
            type: string
            enum: [eu_gdpr, br_lgpd, us_bipa, us_ccpa, other]
        - name: locale
          in: query
          required: false
          schema: { type: string, minLength: 2, maxLength: 16 }
      responses:
        '200':
          description: Active disclosure document.
          content:
            application/json:
              schema:
                type: object
                required: [version, locale, title, bodyMarkdown, jurisdiction]
                properties:
                  version: { type: string }
                  locale: { type: string }
                  title: { type: string }
                  bodyMarkdown: { type: string }
                  jurisdiction: { type: string }
        '400': { $ref: '#/components/responses/BadRequest' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/ServerError' }
...
  /v1/admin/takedowns:
    get:
      operationId: listAdminTakedowns
      tags: [admin]
      summary: List takedown requests (admin:moderate)
      security: [{ bearerAuth: [] }]
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [received, verifying, fulfilled, rejected] }
        - name: overdue
          in: query
          required: false
          schema: { type: boolean }
      responses:
        '200':
          description: Takedown queue.
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      type: object
                      required: [id, subjectEmail, reason, status, slaDueAt, receivedAt]
                      properties:
                        id: { type: string, format: uuid }
                        subjectEmail: { type: string }
                        reason: { type: string }
                        status: { type: string }
                        slaDueAt: { type: string, format: date-time }
                        receivedAt: { type: string, format: date-time }
                        fulfilledAt: { type: string, format: date-time, nullable: true }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/admin/takedowns/{id}/fulfill:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      operationId: fulfillTakedown
      tags: [admin]
      summary: Fulfil a takedown by purging the approved photos (admin:moderate)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [approvedPhotoIds]
              properties:
                approvedPhotoIds:
                  type: array
                  items: { type: string, format: uuid }
                  minItems: 1
                  maxItems: 500
                notes: { type: string, maxLength: 2000 }
      responses:
        '200':
          description: Fulfilment result.
          content:
            application/json:
              schema:
                type: object
                required: [status, fulfilled, failed]
                properties:
                  status: { type: string, enum: [fulfilled] }
                  fulfilled:
                    type: array
                    items: { type: string, format: uuid }
                  failed:
                    type: array
                    items: { type: string, format: uuid }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/ServerError' }
  /v1/admin/takedowns/{id}/reject:
    parameters:
      - $ref: '#/components/parameters/IdParam'
    post:
      operationId: rejectTakedown
      tags: [admin]
      summary: Reject a takedown request (admin:moderate)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rejectionReason]
              properties:
                rejectionReason: { type: string, minLength: 1, maxLength: 2000 }
      responses:
        '200':
          description: Rejection recorded.
          content:
            application/json:
              schema:
                type: object
                required: [status]
                properties:
                  status: { type: string, enum: [rejected] }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '500': { $ref: '#/components/responses/ServerError' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    IdParam:
      in: path
      name: id
      required: true
      schema:
        $ref: '#/components/schemas/Uuid'
...
  responses:
    BadRequest:
      description: Malformed request.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Unauthorized:
      description: Missing or invalid credentials.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    Forbidden:
      description: Authenticated but not permitted.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
...
    Unprocessable:
      description: Validation failed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ServerError:
      description: Unexpected server error.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    Uuid:
      type: string
      format: uuid
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable error code.
            message:
              type: string
            details:
              type: object
              additionalProperties: true
            trace_id:
              type: string
...